refact(api/schemas): migrate workflow-domain local BaseModel classes to schemas_workflows.py (#6042)
Moves 23 classes from 5 workflow endpoint files into schemas_workflows.py:
- workflow.py: WorkflowApprovalResponse, WorkflowApproveResponse, WorkflowCancelResponse,
  WorkflowDetailResponse, WorkflowExecutionRequest, WorkflowExecutionResponse,
  WorkflowListResponse, WorkflowPendingApprovalsResponse, WorkflowStatusResponse, +4 more
- workflow_export.py: ShareWorkflowRequest, ImportWorkflowRequest, CloneWorkflowRequest
- workflow_permissions.py: GrantPermissionRequest, WorkflowPermissionResponse, WorkflowAuditLogEntry
- workflow_secrets.py: WorkflowSecretCreateRequest, WorkflowSecretUpdateRequest, WorkflowSecretMetadata
- workflow_state.py: WorkflowState

Also removes pre-existing duplicate AdvancedControlInfoResponse from schemas_workflows.py.
