+++ autobot-backend/api/workflow_permissions.py
@@ -14,59 +14,28 @@ Routes:
 """
 
 import logging
-from typing import List, Optional
+from typing import List
 
 from fastapi import APIRouter, Depends, HTTPException, Query, status
-from pydantic import BaseModel, Field
 from sqlalchemy.ext.asyncio import AsyncSession
 
+from api.schemas_workflows import (
+    GrantPermissionRequest,
+    WorkflowWorkflowAuditLogEntry,
+    WorkflowWorkflowPermissionResponse,
+)
 from api.user_management.dependencies import get_db_session
 from auth_middleware import get_current_user
+from autobot_shared.error_boundaries import ErrorCategory, with_error_handling
 from services.workflow_permission_service import (
     ROLE_HIERARCHY,
     WorkflowPermissionService,
 )
 from services.workflow_rbac import require_workflow_permission
-from api.schemas_common import DataResponse
-from autobot_shared.error_boundaries import ErrorCategory, with_error_handling
 
 logger = logging.getLogger(__name__)
 
 router = APIRouter(tags=["workflow-permissions"])
-
-
-# ---------------------------------------------------------------------------
-# Request / Response schemas
-# ---------------------------------------------------------------------------
-
-
-class GrantPermissionRequest(BaseModel):
-    """Request body for granting or updating a workflow role."""
-
-    user_id: str = Field(..., description="User receiving the role")
-    role: str = Field(..., description="owner | editor | runner | viewer")
-
-
-class PermissionResponse(BaseModel):
-    """Serialised workflow permission entry."""
-
-    workflow_id: str
-    user_id: str
-    role: str
-    granted_by: Optional[str]
-    created_at: str
-    updated_at: str
-
-
-class AuditLogEntry(BaseModel):
-    """Serialised workflow audit log entry."""
-
-    id: str
-    timestamp: str
-    user_id: str
-    workflow_id: str
-    action: str
-    details: Optional[dict]
 
 
 # ---------------------------------------------------------------------------
@@ -79,9 +48,9 @@ def _current_user_id(current_user: dict) -> str:
     return current_user.get("username") or current_user.get("user_id", "unknown")
 
 
-def _permission_to_response(perm) -> PermissionResponse:
+def _permission_to_response(perm) -> WorkflowWorkflowPermissionResponse:
     """Convert a WorkflowPermission ORM row to its response schema."""
-    return PermissionResponse(
+    return WorkflowWorkflowPermissionResponse(
         workflow_id=perm.workflow_id,
         user_id=perm.user_id,
         role=perm.role,
@@ -91,9 +60,9 @@ def _permission_to_response(perm) -> PermissionResponse:
     )
 
 
-def _audit_to_response(entry) -> AuditLogEntry:
+def _audit_to_response(entry) -> WorkflowWorkflowAuditLogEntry:
     """Convert a WorkflowAuditLog ORM row to its response schema."""
-    return AuditLogEntry(
+    return WorkflowWorkflowAuditLogEntry(
         id=str(entry.id),
         timestamp=entry.timestamp.isoformat() if entry.timestamp else "",
         user_id=entry.user_id,
@@ -115,7 +84,7 @@ def _audit_to_response(entry) -> AuditLogEntry:
 )
 @router.get(
     "/workflows/{workflow_id}/permissions",
-    response_model=List[PermissionResponse],
+    response_model=List[WorkflowPermissionResponse],
     summary="List workflow permissions",
 )
 async def list_workflow_permissions(
@@ -123,7 +92,7 @@ async def list_workflow_permissions(
     current_user: dict = Depends(get_current_user),
     session: AsyncSession = Depends(get_db_session),
     _: bool = Depends(require_workflow_permission("view")),
-) -> List[PermissionResponse]:
+) -> List[WorkflowPermissionResponse]:
     """Return all permission entries for the specified workflow (#2152)."""
     svc = WorkflowPermissionService(session)
     rows = await svc.get_permissions(workflow_id)
@@ -143,7 +112,7 @@ async def list_workflow_permissions(
 )
 @router.put(
     "/workflows/{workflow_id}/permissions",
-    response_model=PermissionResponse,
+    response_model=WorkflowPermissionResponse,
     summary="Grant or update a workflow permission",
 )
 async def grant_workflow_permission(
@@ -152,7 +121,7 @@ async def grant_workflow_permission(
     current_user: dict = Depends(get_current_user),
     session: AsyncSession = Depends(get_db_session),
     _: bool = Depends(require_workflow_permission("grant")),
-) -> PermissionResponse:
+) -> WorkflowPermissionResponse:
     """
     Grant (or update) a role for *body.user_id* on the workflow (#2152).
 
@@ -231,7 +200,7 @@ async def revoke_workflow_permission(
 )
 @router.get(
     "/workflows/{workflow_id}/audit",
-    response_model=List[AuditLogEntry],
+    response_model=List[WorkflowAuditLogEntry],
     summary="Get workflow audit log",
 )
 async def get_workflow_audit_log(
@@ -241,7 +210,7 @@ async def get_workflow_audit_log(
     current_user: dict = Depends(get_current_user),
     session: AsyncSession = Depends(get_db_session),
     _: bool = Depends(require_workflow_permission("view")),
-) -> List[AuditLogEntry]:
+) -> List[WorkflowAuditLogEntry]:
     """Return the audit trail for the specified workflow, newest first (#2152)."""
     svc = WorkflowPermissionService(session)
     entries = await svc.get_audit_log(workflow_id, limit=limit, offset=offset)
